Moved `WindowDimensions` to `window_dimensions.rs`
