+++ wolf_engine_window/src/window.rs
@@ -1,4 +1,4 @@
-use crate::{WindowSettings, FullscreenMode, WindowDimensions};
+use crate::{FullscreenMode, WindowDimensions, WindowSettings};
 
 #[cfg(test)]
 use mockall::automock;
